# src/main/webapp/swagger/matchCenter.json
{
  "swagger": "2.0",
  "info": {
    "description": "赛事平台接口描述，PC和微信都适用",
    "version": "1.0.0",
    "title": "赛事平台接口描述",
    "termsOfService": "http://www.winwayworld.com",
    "contact": {
      "email": "Weijian.he@Winwayworld.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/match",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/login/check": {
      "get": {
        "tags": [
          "login"
        ],
        "summary": "后台登录验证",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "boxFlag",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
...
    "/user/login": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "loginUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The user name for login",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login in clear text",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Result": {
      "type": "object",
      "properties": {
        "code": {
          "type": "number",
          "description": "状态码，0为正确返回"
        },
        "msg": {
          "type": "string",
          "description": "错误信息"
        },
        "detail": {
          "type": "string",
          "description": "错误详细描述或返回码对应处理方案"
        },
        "data": {
          "type": "object",
          "description": "返回数据"
        }
      }
    },
...
  }
}